whitespace
I'm sorry for making whitespace-only changes. Please feel free to revert if it causes any merge problems whatsoever.

Let's come up with a style that we can all be happy with that makes the code easy to read/comprehend without being too pedantic.

We should be able to set our IDEs to the same settings, so it's something we don't have think about.

We should at least being consistent with respect to indentation depth for methods within classes, blocks of code with curly-braces and so on.