A general round of cleanup and refactoring in PGSecondary and related classes.
- pass the PGSecondaryUniverse around instead of the Connection.
The object lifecycle needs a bit more work here but I think this
is a step in the right direction.

- refactor code to lookup the truthCopyInfo in PGSecondary so we don't
have to duplicate that code all over.

- handle WorkingCopyCreated event up front to simplify code path.  I may
want to refactor the specifics of how this is done again, scala
Iterators aren't giving me quite what I want yet.

- pull out remaining events into event handlers.

- refactor how RowDataUpdated is handled, making it more efficient and
also adding support for updates and deletes.