Add support for the "_" column name prefix that ensures column names are valid soql.
soda-fountain will now add a "_" prefix to non-system columns to ensure the soql generated
is valid (ie. column id doesn't start with a number), so we have to deal with it when parsing
the query.  This consists of adding it to the dataset schema before parsing the query, then
removing it from the parsed query before querying the dataset.