Create fewer indexes for rollups.
Text indexes are by _far_ the slowest to create due to having to
deal with collations.  This should approximately halve the index
creation time for text columns.  We are creating 1/3rd of the indexes,
but the text_pattern_ops one was cheaper since it didn't have the big
hit from collation support.

As per the comment in the code:

  In the bigger picture, the choice of indexes to create should be more dynamic based on
  type of queries we expect to the table and/or dynamically determined from workload.
  For now we are putting minimal indexes on rollup tables since their main use case is
  making small tables we can efficiently do full table scans on.  We are not typically
  doing prefix matching or case insensitive matching against them.