EN-4323: Implement Secondary.dropCopy(.)
Secondary.dropCopy(.) is part of the resync
path for discarded and snappshotted copies.
Keep an entry in `copy_map` for a discarded
copy during resync only if it is the latest
copy of the dataset.

We want to do this because secondary-watcher
updates the secondary manifest to say that
a secondary is done with the _latest_ copy
of a dataset after resync.