Handle changes to the data-coordinator library
Handle breaking changes in the data-coordinator/secondary
library version 3.3.7. Includes work for adding resource name to
truth so the secondary-watchers can eventually send replication
messages on AMQ containing a dataset's view uid (EN-15144):

 - DatasetInfo classes now have an optional field resourceName
 - `dataset_map` table now has a column `resource_name`

Adds a migration to add the `resource_name` column to `dataset_map`.
This migration will require the secondary-watchers to be off when run.