EN-16779: Enable parallel queries by setting fetchSize to 0
As per https://www.postgresql.org/docs/9.6/static/when-can-parallel-query-be-used.html parallel
query can not be used if:

  The client sends an Execute message with a non-zero fetch count.
  See the discussion of the extended query protocol. Since libpq
  currently provides no way to send such a message, this can only
  occur when using a client that does not rely on libpq. If this is
  a frequent occurrence, it may be a good idea to set
  max_parallel_workers_per_gather in sessions where it is likely, so
  as to avoid generating query plans that may be suboptimal when run
  serially.

We set a fetch size so we can stream resuls without the jdbc driver
buffering them all.  This results in the driver using a cursor which
prevents parallel query action.

However, given that most of the queries that really benefit from
parallel queries are aggregation queries, and we have a default
limit of 1000 rows returned (which just happens to be the same as
the fetch size we were using), if we know the query has a limit <=
our desired fetch size then we can just set it to 0 without worrying
about running out of memory, and let postgres use parallel queries.