Change continuous median() so it works as a window function.
For now we are only doing this for the continuous median as that
is where our current use cases lie.

This implementation can be significantly slower.  Ideally we would
know if we were using it inside a window function or not so we could
switch implementations but that would take some more work.