Allow "approximatelyEqual" to be _even more_ approximate
This gives us a place we can hook into the comparison project within
the fixtures themselves to say "this particular sub-value should be
considered good given some arbitrary condition".

It can be extended by creating new subclasses of ExpectedSpec and
implementing `accept`.  When using them in a fixture, create an
object with a key `expectation` that names the type of expectation.
Other parameters in that object are used to provide parameters of
the subclass.

Right now, the only such subclass is one that expects that the
corresponding JSON value be a base64 string.