Don't compile `get_context(stringlit)` to a current_setting call
No need for it in that case, we can just look the value up in the map
ourselves and put it in the SQL as if it had been a string literal
itself.

If a non-string-literal is passed in, we'll still do the runtime lookup
but this should basically never happen.