EN-60548: Fake-locations on the new query path
Plumb the subcolumn info from upstream into the SoQLRepProvider
so that SoQLLocation can select them into the normal compound column
intermediate format.

The "specialness" of this is pretty nicely kept to a minimum - the
subcolumn info needs to get passed into the soql rep provider, and
we need to keep track of what table labels refer to what physical
database tables, but other than that all the weirdness of fake-locations
is managed entirely within SoQLLocation's Rep implementation.