Stream results from the database
Postgresql's JDBC driver will read rows in fixed-size blocks; by default,
it reads _all_ the results into memory before returning them.  You can
make it stream by setting the fetch size (in rows), so we'll do that.

Specifically, we'll set the fetch size based on the types of columns in
the result - in particular geo types routinely use a lot of memory, so
we'll consider such types as being "large" and set a small fetch size
when one exists in the result.