EN-62111: type-annotate number literals in the old sqlizer
When the old sqlizer would turn a number literal into a SQL number
literal, it didn't annotate it as any particular type.  As a result,
if your number literal happened to stringify as an integer, further
math operations on it would be integer math operations.  As a result,
for example, `select 1/4` would return 0 instead of 0.25 as would be
expected.

Now instead of sqlizing as something like `1/4` that expression will
sqlize as something like `(1::numeric)/(4::numeric)`.

New-sqlizer already does this