Only create FTS indexes on rollup if they look like materialized views
One of the things you can do in this new system is say "create a rollup
with the query of this saved view" and while this doesn't imply
automatically regenerating this rollup when that saved view changes, the
fact that it was initially created from a view is propagated down to the
place it's turned into a SQL table.  So, since FTS indexes are easily
the slowest kind of index to create, let's only do that if it looks like
the "rollup" is actually a "materialized view", since otherwise it's
extremely unlikely to ever actually be used.