Don't use pg config for dynamic system context lookups
This was always a hack; PG's settings system was absolutely not meant
for this and it's always bothered me that we were using it even beyond
the bug that if you happened to use an unset config val in a session
where it had been previously set, you'd get back an empty string insetad
of null.

Now instead if dynamic config vals are used in a query, we create a
function in pg_temp (which is specifically for storing per-transaction
stuff) and call it there.