ECS shutdown experiment
When I was doing the Redis thing last week, I noticed that we're still
apparently not shutting down properly in ECS.  Best guess is that sudo
is eating the SIGTERM that ECS is supposed to send us when a shutdown
is initiated.  So, let's see if that's the case by _not_ switching user
if we're ECSified.